updated put and delete api's

+++ routes/achievements.js
@@ -4,7 +4,7 @@ const router = express.Router();
 
 const { Achievement, validation } = require("../models/achievements");
 const { User } = require("../models/user");
-
+// Create
 router.post("/:email", async (req, res) => {
   const { error } = validation.validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
@@ -13,8 +13,16 @@ router.post("/:email", async (req, res) => {
   if (!user)
     return res.status(400).send("User with this email is not registered.");
 
+  user = await User.findById(user._id).populate("achievements");
+  let achievement = user.achievements;
+  achievement = achievement.filter(
+    (obj) => obj.achievement_id === req.body.achievement_id
+  );
+  if (achievement.length)
+    return res.status(400).send("Achievement with this id is already added");
+
   achievement = new Achievement(
-    _.pick(req.body, ["achievement", "description", "year"])
+    _.pick(req.body, ["achievement_id", "achievement", "description", "year"])
   );
   await achievement.save();
 
@@ -23,9 +31,17 @@ router.post("/:email", async (req, res) => {
     { $push: { achievements: achievement._id } }
   );
 
-  res.send(_.pick(achievement, ["_id", "achievement", "description", "year"]));
+  res.send(
+    _.pick(achievement, [
+      "_id",
+      "achievement_id",
+      "achievement",
+      "description",
+      "year",
+    ])
+  );
 });
-
+// Read
 router.get("/get/:email", async (req, res) => {
   //const achievements = await Achievement.find().sort({ year: 1 });
   let user = await User.findOne({ email: req.params.email });
@@ -36,26 +52,64 @@ router.get("/get/:email", async (req, res) => {
   // Sort below achievement.achievements object as it is not sorted yet
   res.send(achievement.achievements);
 });
+// delete
+router.delete("/delete/:email/:id", async (req, res) => {
+  let user = await User.findOne({ email: req.params.email });
+  if (!user)
+    return res.status(400).send("User with this email is not registered.");
 
-router.delete("/:id", async (req, res) => {
-  let achievement = await Achievement.findOneAndRemove({
-    _id: req.params.id,
+  user = await User.findById(user._id).populate("achievements");
+  let achievement = user.achievements;
+  let id_to_delete = achievement
+    .map((obj) => {
+      if (obj.achievement_id === parseInt(req.params.id)) return obj._id;
+      else return undefined;
+    })
+    .filter((obj) => obj !== undefined);
+
+  if (id_to_delete.length === 0)
+    return res.status(400).send("This Achievement is not added yet.");
+
+  let updatedAchievements = achievement.filter(
+    (obj) => obj.achievement_id !== parseInt(req.params.id)
+  );
+  await User.updateOne(
+    { _id: user._id },
+    { $set: { achievements: updatedAchievements } }
+  );
+  await Achievement.findOneAndRemove({
+    _id: id_to_delete,
   });
 
-  res.send(achievement);
+  res.send("Deleted Succesfully");
 });
-router.put("/:id", async (req, res) => {
+// update
+router.put("/update/:email/:id", async (req, res) => {
+  let user = await User.findOne({ email: req.params.email });
+  if (!user)
+    return res.status(400).send("User with this email is not registered.");
+
+  user = await User.findById(user._id).populate("achievements");
+  let achievement = user.achievements;
+  let id_to_update = achievement
+    .map((obj) => {
+      if (obj.achievement_id === parseInt(req.params.id)) return obj._id;
+      else return undefined;
+    })
+    .filter((obj) => obj !== undefined);
+
+  if (id_to_update.length === 0)
+    return res.status(400).send("This Achievement is not added yet.");
+
   const { error } = validation.validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
-  let achievement = await Achievement.findOneAndUpdate({
-    _id: req.params.id,
-    achievement: req.body.achievement,
-    description: req.body.description,
-    year: req.body.year,
-  });
+  await Achievement.findOneAndUpdate(
+    { _id: id_to_update, achievement_id: parseInt(req.params.id) },
+    req.body
+  );
 
-  res.send(achievement);
+  res.send("Updated Succesfully");
 });
 
 module.exports = router;
