+++ routes/sportInfo.js
@@ -12,8 +12,14 @@ router.post("/:email", async (req, res) => {
   if (!user)
     return res.status(400).send("User with this email is not registered.");
 
+  user = await User.findById(user._id).populate("sportInfo");
+  let sport = user.sportInfo;
+  sport = sport.filter((obj) => obj.sport_id === req.body.sport_id);
+  if (sport.length)
+    return res.status(400).send("Sport with this id is already added");
+
   let info = new SportInfo(
-    _.pick(req.body, ["sport", "teamName", "position_in_team"])
+    _.pick(req.body, ["sport_id", "sport", "teamName", "position_in_team"])
   );
   await info.save();
 
@@ -21,7 +27,9 @@ router.post("/:email", async (req, res) => {
     { email: req.params.email },
     { $push: { sportInfo: info._id } }
   );
-  res.send(_.pick(info, ["_id", "sport", "teamName", "position_in_team"]));
+  res.send(
+    _.pick(info, ["_id", "sport_id", "sport", "teamName", "position_in_team"])
+  );
 });
 
 router.get("/get/:email", async (req, res) => {
@@ -32,25 +40,61 @@ router.get("/get/:email", async (req, res) => {
   res.send(sportInfo.sportInfo);
 });
 
-router.delete("/:id", async (req, res) => {
-  let info = await SportInfo.findOneAndRemove({
-    _id: req.params.id,
+router.delete("/delete/:email/:id", async (req, res) => {
+  let user = await User.findOne({ email: req.params.email });
+  if (!user)
+    return res.status(400).send("User with this email is not registered.");
+
+  user = await User.findById(user._id).populate("sportInfo");
+  let sport = user.sportInfo;
+  let id_to_delete = sport
+    .map((obj) => {
+      if (obj.sport_id === parseInt(req.params.id)) return obj._id;
+      else return undefined;
+    })
+    .filter((obj) => obj !== undefined);
+
+  if (id_to_delete.length === 0)
+    return res.status(400).send("This Sport Information is not added yet.");
+  let updatedSports = sport.filter(
+    (obj) => obj.sport_id !== parseInt(req.params.id)
+  );
+  await User.updateOne(
+    { _id: user._id },
+    { $set: { sportInfo: updatedSports } }
+  );
+  await SportInfo.findOneAndRemove({
+    _id: id_to_delete,
   });
 
-  res.send(info);
+  res.send("Deleted Succesfully");
 });
-router.put("/:id", async (req, res) => {
+router.put("/update/:email/:id", async (req, res) => {
+  let user = await User.findOne({ email: req.params.email });
+  if (!user)
+    return res.status(400).send("User with this email is not registered.");
+
+  user = await User.findById(user._id).populate("sportInfo");
+  let sport = user.sportInfo;
+  let id_to_update = sport
+    .map((obj) => {
+      if (obj.sport_id === parseInt(req.params.id)) return obj._id;
+      else return undefined;
+    })
+    .filter((obj) => obj !== undefined);
+
+  if (id_to_update.length === 0)
+    return res.status(400).send("This Sport Information is not added yet.");
+
   const { error } = validation.validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
-  let info = await SportInfo.findOneAndUpdate({
-    _id: req.params.id,
-    sport: req.body.sport,
-    teamName: req.body.teamName,
-    position_in_team: req.body.position_in_team,
-  });
+  await SportInfo.findOneAndUpdate(
+    { _id: id_to_update, sport_id: parseInt(req.params.id) },
+    req.body
+  );
 
-  res.send(info);
+  res.send("Updated Succesfully");
 });
 
 module.exports = router;
