+++ routes/statistics.js
@@ -11,9 +11,17 @@ router.post("/:email", async (req, res) => {
   let user = await User.findOne({ email: req.params.email });
   if (!user)
     return res.status(400).send("User with this email is not registered.");
+  user = await User.findById(user._id).populate("statistics");
+  let statistic = user.statistics;
+  statistic = statistic.filter(
+    (obj) => obj.statistic_id === req.body.statistic_id
+  );
+  if (statistic.length)
+    return res.status(400).send("Statistic with this id is already added");
 
-  let statistic = new Statistic(
+  statistic = new Statistic(
     _.pick(req.body, [
+      "statistic_id",
       "tournament",
       "total_matches",
       "total_score",
@@ -32,6 +40,7 @@ router.post("/:email", async (req, res) => {
   res.send(
     _.pick(statistic, [
       "_id",
+      "statistic_id",
       "tournament",
       "total_matches",
       "total_score",
@@ -51,28 +60,60 @@ router.get("/get/:email", async (req, res) => {
   res.send(statistics.statistics);
 });
 
-router.delete("/:id", async (req, res) => {
-  let statistic = await Statistic.findOneAndRemove({
-    _id: req.params.id,
+router.delete("/delete/:email/:id", async (req, res) => {
+  let user = await User.findOne({ email: req.params.email });
+  if (!user)
+    return res.status(400).send("User with this email is not registered.");
+
+  user = await User.findById(user._id).populate("statistics");
+  let statistic = user.statistics;
+  let id_to_delete = statistic
+    .map((obj) => {
+      if (obj.statistic_id === parseInt(req.params.id)) return obj._id;
+      else return undefined;
+    })
+    .filter((obj) => obj !== undefined);
+
+  if (id_to_delete.length === 0)
+    return res.status(400).send("This Statistic is not added yet.");
+  let updatedStatistics = statistic.filter(
+    (obj) => obj.statistic_id !== parseInt(req.params.id)
+  );
+  await User.updateOne(
+    { _id: user._id },
+    { $set: { statistics: updatedStatistics } }
+  );
+  await Statistic.findOneAndRemove({
+    _id: id_to_delete,
   });
 
-  res.send(statistic);
+  res.send("Deleted Succesfully");
 });
-router.put("/:id", async (req, res) => {
+router.put("/update/:email/:id", async (req, res) => {
+  let user = await User.findOne({ email: req.params.email });
+  if (!user)
+    return res.status(400).send("User with this email is not registered.");
+
+  user = await User.findById(user._id).populate("statistics");
+  let statistic = user.statistics;
+  let id_to_update = statistic
+    .map((obj) => {
+      if (obj.statistic_id === parseInt(req.params.id)) return obj._id;
+      else return undefined;
+    })
+    .filter((obj) => obj !== undefined);
+
+  if (id_to_update.length === 0)
+    return res.status(400).send("This Statistic is not added yet.");
   const { error } = validation.validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
-  let statistic = await Statistic.findOneAndUpdate({
-    _id: req.params.id,
-    tournament: req.body.tournament,
-    total_matches: req.body.total_matches,
-    total_score: req.body.total_score,
-    total_wickets: req.body.total_wickets,
-    club: req.body.club,
-    total_goals: req.body.total_goals,
-  });
+  await Statistic.findOneAndUpdate(
+    { _id: id_to_update, statistic_id: parseInt(req.params.id) },
+    req.body
+  );
 
-  res.send(statistic);
+  res.send("Updated Succesfully");
 });
 
 module.exports = router;
